cleaned up OpenCL code some more and moved the OpenCL helper functions to their own file. Additionally, added more specific timing data for program runs
